requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.1.0
Markdown==3.5.1